fix properlike
