+++ src/core/lib/slice/slice_buffer.cc
@@ -411,50 +411,3 @@ void grpc_slice_buffer_undo_take_first(grpc_slice_buffer* sb,
   sb->count++;
   sb->length += GRPC_SLICE_LENGTH(slice);
 }
-
-////////////////////////////////////////////////////////////////////////////////
-// TODO(ctiller): the rest of this file is refugees from slice_string_helpers.cc
-// that should be moved into a better place.
-
-static void grpc_slice_split_inner(grpc_slice str, const char* sep,
-                                   grpc_slice_buffer* dst, bool no_space) {
-  const size_t sep_len = strlen(sep);
-  size_t begin, end;
-  const uint8_t* str_buffer = GRPC_SLICE_START_PTR(str);
-  size_t sep_pos;
-
-  GPR_ASSERT(sep_len > 0);
-
-  if (slice_find_separator_offset(str, sep, 0, &begin, &end) != 0) {
-    do {
-      sep_pos = end;
-      if (no_space) {
-        skip_leading_trailing_spaces(str_buffer, &begin, &end);
-      }
-      grpc_slice_buffer_add_indexed(dst, grpc_slice_sub(str, begin, end));
-    } while (slice_find_separator_offset(str, sep, sep_pos + sep_len, &begin,
-                                         &end) != 0);
-    begin = sep_pos + sep_len;
-    end = GRPC_SLICE_LENGTH(str);
-    if (no_space) {
-      skip_leading_trailing_spaces(str_buffer, &begin, &end);
-    }
-    grpc_slice_buffer_add_indexed(dst, grpc_slice_sub(str, begin, end));
-  } else { /* no sep found, add whole input */
-    begin = 0;
-    end = GRPC_SLICE_LENGTH(str);
-    if (no_space) {
-      skip_leading_trailing_spaces(str_buffer, &begin, &end);
-    }
-    grpc_slice_buffer_add_indexed(dst, grpc_slice_sub(str, begin, end));
-  }
-}
-
-void grpc_slice_split(grpc_slice str, const char* sep, grpc_slice_buffer* dst) {
-  grpc_slice_split_inner(str, sep, dst, false);
-}
-
-void grpc_slice_split_without_space(grpc_slice str, const char* sep,
-                                    grpc_slice_buffer* dst) {
-  grpc_slice_split_inner(str, sep, dst, true);
-}
