+++ src/core/lib/slice/slice_string_helpers.cc
@@ -20,10 +20,6 @@
 
 #include "src/core/lib/slice/slice_string_helpers.h"
 
-#include <string.h>
-
-#include <grpc/support/log.h>
-
 #include "src/core/lib/gpr/string.h"
 #include "src/core/lib/gprpp/memory.h"
 #include "src/core/lib/slice/slice_internal.h"
@@ -39,42 +35,6 @@ grpc_slice grpc_dump_slice_to_slice(const grpc_slice& s, uint32_t flags) {
       gpr_dump_return_len(reinterpret_cast<const char*> GRPC_SLICE_START_PTR(s),
                           GRPC_SLICE_LENGTH(s), flags, &len));
   return grpc_slice_from_moved_buffer(std::move(ptr), len);
-}
-
-/** Finds the initial (\a begin) and final (\a end) offsets of the next
- * substring from \a str + \a read_offset until the next \a sep or the end of \a
- * str.
- *
- * Returns 1 and updates \a begin and \a end. Returns 0 otherwise. */
-static int slice_find_separator_offset(const grpc_slice str, const char* sep,
-                                       const size_t read_offset, size_t* begin,
-                                       size_t* end) {
-  size_t i;
-  const uint8_t* str_ptr = GRPC_SLICE_START_PTR(str) + read_offset;
-  const size_t str_len = GRPC_SLICE_LENGTH(str) - read_offset;
-  const size_t sep_len = strlen(sep);
-  if (str_len < sep_len) {
-    return 0;
-  }
-
-  for (i = 0; i <= str_len - sep_len; i++) {
-    if (memcmp(str_ptr + i, sep, sep_len) == 0) {
-      *begin = read_offset;
-      *end = read_offset + i;
-      return 1;
-    }
-  }
-  return 0;
-}
-
-static void skip_leading_trailing_spaces(const uint8_t* str_buffer,
-                                         size_t* begin, size_t* end) {
-  while (*begin < *end && str_buffer[*begin] == ' ') {
-    (*begin)++;
-  }
-  while (*begin < *end && str_buffer[*end - 1] == ' ') {
-    (*end)--;
-  }
 }
 
 bool grpc_parse_slice_to_uint32(grpc_slice str, uint32_t* result) {
