+++ test/core/slice/BUILD
@@ -87,6 +87,18 @@ grpc_cc_test(
     uses_polling = False,
     deps = [
         "//:gpr",
+        "//:slice",
+        "//test/core/util:grpc_suppressions",
+    ],
+)
+
+grpc_cc_test(
+    name = "slice_intern_test",
+    srcs = ["slice_intern_test.cc"],
+    language = "C++",
+    uses_polling = False,
+    deps = [
+        "//:gpr",
         "//:grpc",
         "//test/core/util:grpc_test_util",
     ],
@@ -95,6 +107,18 @@ grpc_cc_test(
 grpc_cc_test(
     name = "slice_string_helpers_test",
     srcs = ["slice_string_helpers_test.cc"],
+    language = "C++",
+    uses_polling = False,
+    deps = [
+        "//:gpr",
+        "//:slice",
+        "//test/core/util:grpc_suppressions",
+    ],
+)
+
+grpc_cc_test(
+    name = "slice_split_test",
+    srcs = ["slice_split_test.cc"],
     language = "C++",
     uses_polling = False,
     deps = [
